Move `ioHasFS` related modules into an internal hierachy.
