+++ fs-api/fs-api.cabal
@@ -34,9 +34,9 @@ library
     System.FS.API
     System.FS.API.Types
     System.FS.CRC
-    System.FS.Handle
     System.FS.IO
-    System.IO.FS
+    System.FS.IO.Internal
+    System.FS.IO.Internal.Handle
     Util.CallStack
     Util.Condense
