+++ fs-api/src-win32/System/FS/IO/Internal.hs
@@ -1,5 +1,8 @@
 {-# OPTIONS_GHC -Wno-dodgy-imports #-}
-module System.IO.FS (
+
+-- | This is meant to be used for the implementation of HasFS instances and not
+-- directly by client code.
+module System.FS.IO.Internal (
     FHandle
   , close
   , getSize
@@ -22,7 +25,7 @@ import           Data.Word (Word32, Word64, Word8)
 import           Foreign (Int64, Ptr)
 import           System.FS.API.Types (AllowExisting (..), FsError (..),
                      FsErrorType (..), OpenMode (..), SeekMode (..))
-import           System.FS.Handle
+import           System.FS.IO.Internal.Handle
 import           System.Win32
 
 type FHandle = HandleOS HANDLE
